+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Bugs test suite
+### Fixed
+- Fix `No free channel id` message when emit large amount of events
 
 ## [1.2.0-beta.2]
 ### Fixed